requirements.txt
aiogram
langchain
langchain-openai
langgraph
langgraph-checkpoint-sqlite
python-dotenv
requests